E14093051_GAI_Project2b/requirements.txt
absl-py==2.1.0
accelerate==0.29.1
aiofiles==22.1.0
aiohttp==3.9.3
aioquic==0.9.20
aiosignal==1.3.1
anyio==3.6.1
appdirs==1.4.4
asgiref==3.5.2
async-timeout==4.0.3
attrs==23.2.0
blinker==1.5
brotlipy==0.7.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.9
click==8.1.3
colorama==0.4.5
contourpy==1.2.1
cpm-kernels==1.0.11
cryptography==38.0.1
cycler==0.12.1
datasets==2.18.0
decorator==5.1.1
dill==0.3.8
dnslib==0.9.22
docker-pycreds==0.4.0
filelock==3.13.3
Flask==2.1.3
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.2.0
future==0.18.3
gensim==4.3.2
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.62.1
h11==0.14.0
httpbin==0.7.0
huggingface-hub==0.22.2
idna==3.3
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
kociemba==1.2.1
loguru==0.7.2
Markdown==3.6
MarkupSafe==2.1.1
matplotlib==3.8.4
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
numpy==1.26.1
OpenCC==1.1.7
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
packaging==24.0
pandas==2.1.3
peft==0.10.0
Pillow==10.0.0
protobuf==4.25.3
psutil==5.9.8
pyarrow==15.0.2
pyarrow-hotfix==0.6
pycparser==2.21
pylsqpack==0.3.16
pyOpenSSL==22.1.0
pyparsing==3.1.2
python-dateutil==2.8.2
pytorch-ignite==0.5.0.post2
pytz==2021.3
PyYAML==6.0.1
raven==6.10.0
regex==2023.12.25
requests==2.26.0
rouge==1.0.1
rouge-chinese==1.0.3
sacremoses==0.1.1
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.11.3
sentencepiece==0.2.0
sentry-sdk==1.44.1
setproctitle==1.3.3
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.0
starlette==0.21.0
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
text2vec==1.2.9
textgen==1.1.1
threadpoolctl==3.2.0
tokenizers==0.15.2
torch==2.2.2+cu121
torchaudio==2.2.2+cu121
torcheval==0.0.7
torchtext==0.17.2
torchvision==0.17.2+cu121
tqdm==4.66.2
transformers==4.39.3
typing_extensions==4.11.0
tzdata==2023.3
urllib3==1.26.18
wandb==0.16.6
Werkzeug==2.0.3
win32-setctime==1.1.0
wrapt==1.16.0
wsproto==1.2.0
xformers @ https://github.com/C43H66N12O12S2/stable-diffusion-webui/releases/download/f/xformers-0.0.14.dev0-cp310-cp310-win_amd64.whl
xxhash==3.4.1
yarl==1.9.4
